Sink/Souce was backwards. Swapped. Added some sensible constants to deal with invert oddities on the p channel mosfets. Re-enabled PID speed control, and encoder sensor. PID works well, pos sensor less so.

+++ src/motor/motor_controller.h
@@ -20,7 +20,7 @@ class MotorController
 private:
 	MotorSensors Sensors;
 	MotorDriver Motor;
-	PIDController<float, uint8_t> SpeedPID;
+	PIDController<float, int> SpeedPID;
 	
 	uint8_t TargetRPM{0};
